Create home component
